requirements.txt
imapclient
beautifulsoup4
python-dotenv
